fix(store): make batch outcome counts intrinsic and harmonize error mappings
- Carry intrinsic counts in BatchLandingVerdict variants (LandedAll, PreAttemptRefusal, PartialProgress) so helpers derive counts directly without caller totals or min clamps (resolves M4).
- Add discriminating suffix partition guard test test_m4_inconsistent_suffix_partition_rejected (resolves M11).
- Map single-event NATS publish initiation error to TransportUnavailable matching batch (resolves M1).
- Compare fault limits in u64 domain without narrowing cast in FileEngine (resolves L6).
- Rename file sync test to reflect sequential stop-before-suffix contract (resolves R7-L1).

+++ crates/pardosa/src/file.rs
@@ -1203,7 +1203,7 @@ impl StorageEngine for FileEngine {
         }
 
         if let Some(limit) = self.undetermined_after_n_blocks {
-            if (self.frame_count as usize) >= limit {
+            if self.frame_count >= limit as u64 {
                 self.uncertain = true;
                 self.uncertain_diagnostic = Some(
                     "write landing undetermined: simulated indeterminate write landing".to_string(),
@@ -1215,7 +1215,7 @@ impl StorageEngine for FileEngine {
         }
 
         if let Some(limit) = self.fail_after_n_blocks {
-            if (self.frame_count as usize) >= limit {
+            if self.frame_count >= limit as u64 {
                 return Err(OperationFailure::new(
                     FailureCondition::PrecursorChainBroken(None),
                     "simulated write failure after limit",
